fix(emit): preserve invalid let for recovery
