+++ crates/tsz-emitter/src/emitter/statements/mod.rs
@@ -2,10 +2,12 @@ mod async_super_capture;
 mod control_flow;
 mod core;
 mod expression_statement_helpers;
+mod for_recovery;
 mod recovered_expression_statement_helpers;
 mod recovered_generated_member_helpers;
 mod static_block_await_recovery;
 mod try_statement;
+mod using_helpers;
 mod variable_statement_helpers;
 
 #[cfg(test)]
